requirements.txt
alembic==1.7.7
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
DateTime==4.4
dnspython==2.2.1
docutils==0.18.1
elastic-transport==8.1.1
email-validator==1.1.3
Flask==2.0.3
Flask-APScheduler==1.12.3
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.32
statistics==1.0.3.5
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
Werkzeug==2.0.3
WTForms==3.0.1
zope.interface==5.4.0
